requirements-dev.txt
SQLAlchemy
beautifulsoup4
bitstring
bottle
cffi
dnspython
enum34
flake8
idna
ipaddress
pbkdf2
py3kwarn
pyasn1
pycparser
pycrypto
pyelftools
pylzma
pypdns
pypssl
pytest
pytest-cov
pytest-django
python-dateutil
python-magic
r2pipe
requests-cache
scandir
six
terminaltables
tox
virustotal-api
yara-python
cryptography
requests[socks]

git+https://github.com/viper-framework/pefile.git#egg=pefile
git+https://github.com/smarnach/pyexiftool.git#egg=pyexiftool
git+https://github.com/kbandla/pydeep.git#egg=pydeep

# pe module (verify sigs)
git+https://github.com/sebdraven/verify-sigs.git

# office module
olefile
git+https://github.com/decalage2/oletools.git
## Flash lookup
git+https://github.com/viper-framework/xxxswf.git


# clamd module
pyclamd

# apk module
ipython < 6; python_version == '2.7'
git+https://github.com/androguard/androguard.git

# archiving
rarfile

# MISP
git+https://github.com/MISP/PyTaxonomies.git
git+https://github.com/MISP/PyMISPGalaxies.git
pymisp[fileobjects,virustotal] >= 2.4.84

# Scraper
git+https://github.com/viper-framework/ScrapySplashWrapper.git

# Joe Sandbox
jbxapi
